python-iot-gateway/requirements.txt
kafka-python-ng==2.2.2
pymongo==4.10.1
